refactor(desktop): extract FileViewerPane into smaller components and consolidate Monaco utilities
