+++ apps/desktop/src/lib/trpc/routers/changes/utils/parse-status.test.ts
@@ -1,10 +1,6 @@
 import { describe, expect, test } from "bun:test";
-import {
-	detectLanguage,
-	parseDiffNumstat,
-	parseGitLog,
-	parseNameStatus,
-} from "./parse-status";
+import { detectLanguage } from "shared/detect-language";
+import { parseDiffNumstat, parseGitLog, parseNameStatus } from "./parse-status";
 
 describe("parseGitLog", () => {
 	test("parses basic log output", () => {
