+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/FileViewerPane/components/FileViewerContent/index.ts
@@ -1,0 +1,1 @@
+export { FileViewerContent } from "./FileViewerContent";
