+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/FileViewerPane/components/FileViewerToolbar/index.ts
@@ -1,0 +1,1 @@
+export { FileViewerToolbar } from "./FileViewerToolbar";
