+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/FileViewerPane/hooks/useFileContent/index.ts
@@ -1,0 +1,1 @@
+export { useFileContent } from "./useFileContent";
