+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/FileViewerPane/hooks/useSplitOrientation/index.ts
@@ -1,0 +1,4 @@
+export {
+	type SplitOrientation,
+	useSplitOrientation,
+} from "./useSplitOrientation";
